schema_bench/requirements.txt
pydantic
jsonschema
pyyaml
openai
python-dotenv
